+++ Class/Class.Token/Info.cs
@@ -2,6 +2,6 @@ namespace Class.Token;
 
 public class Info : Any
 {
-    public virtual int Row { get; set; }
+    public virtual long Row { get; set; }
     public virtual Range Range { get; set; }
 }
